requirements.txt
Django>=4.2
djangorestframework>=3.14
Pillow>=10.0
numpy>=1.24
django-cors-headers>=4.2

# Production deployment
gunicorn>=21.2
whitenoise>=6.5
dj-database-url>=2.1
psycopg2-binary>=2.9
python-dotenv>=1.0

# Optional: Face recognition (may have build issues on some platforms)
# opencv-python>=4.8
# opencv-contrib-python>=4.8
# deepface>=0.0.79
# mtcnn>=0.1.1

# Development only
django-extensions>=3.2
Werkzeug>=3.0
pyOpenSSL>=23.0
